requirements.txt
fastapi
uvicorn
python-dotenv
pydantic-settings
loguru
aiogram
python-multipart
python-telegram-bot